stock/serializers.py
# stock/serializers.py

from rest_framework import serializers
from .models import Article

class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = '__all__'  # Utilisez '__all__' pour inclure tous les champs du modèle
