requirements.txt
Django==4.2.5
djangorestframework==3.12.4
djangorestframework-authtoken==2.1.4
